+++ pydoclint/main.py
@@ -1,8 +1,9 @@
+from __future__ import annotations
+
 import ast
 import logging
 import re
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
 
 import click
 
@@ -28,8 +29,8 @@ echoAsError = True
 def validateStyleValue(
         context: click.Context,
         param: click.Parameter,
-        value: Optional[str],
-) -> Optional[str]:
+        value: str | None,
+) -> str | None:
     """Validate the value of the 'style' option"""
     if value not in {'numpy', 'google', 'sphinx'}:
         raise click.BadParameter(
@@ -306,7 +307,7 @@ def main(  # noqa: C901
         quiet: bool,
         exclude: str,
         style: str,
-        paths: Tuple[str, ...],
+        paths: tuple[str, ...],
         type_hints_in_signature: str,
         type_hints_in_docstring: str,
         arg_type_hints_in_signature: bool,
@@ -327,7 +328,7 @@ def main(  # noqa: C901
         generate_baseline: bool,
         baseline: str,
         show_filenames_in_every_violation_message: bool,
-        config: Optional[str],  # don't remove it b/c it's required by `click`
+        config: str | None,  # don't remove it b/c it's required by `click`
 ) -> None:
     """Command-line entry point of pydoclint"""
     logging.basicConfig(level=logging.WARN if quiet else logging.INFO)
@@ -392,7 +393,7 @@ def main(  # noqa: C901
             )
             ctx.exit(1)
 
-    violationsInAllFiles: Dict[str, List[Violation]] = _checkPaths(
+    violationsInAllFiles: dict[str, list[Violation]] = _checkPaths(
         quiet=quiet,
         exclude=exclude,
         style=style,
@@ -516,7 +517,7 @@ def main(  # noqa: C901
 
 
 def _checkPaths(
-        paths: Tuple[str, ...],
+        paths: tuple[str, ...],
         style: str = 'numpy',
         argTypeHintsInSignature: bool = True,
         argTypeHintsInDocstring: bool = True,
@@ -534,8 +535,8 @@ def _checkPaths(
         requireYieldSectionWhenYieldingNothing: bool = False,
         quiet: bool = False,
         exclude: str = '',
-) -> Dict[str, List[Violation]]:
-    filenames: List[Path] = []
+) -> dict[str, list[Violation]]:
+    filenames: list[Path] = []
 
     if not quiet:
         skipMsg = f'Skipping files that match this pattern: {exclude}'
@@ -552,7 +553,7 @@ def _checkPaths(
         elif path.is_dir():
             filenames.extend(sorted(path.rglob('*.py')))
 
-    allViolations: Dict[str, List[Violation]] = {}
+    allViolations: dict[str, list[Violation]] = {}
 
     for filename in filenames:
         if excludePattern.search(filename.as_posix()):
@@ -563,7 +564,7 @@ def _checkPaths(
                 click.style(filename, fg='cyan', bold=True), err=echoAsError
             )
 
-        violationsInThisFile: List[Violation] = _checkFile(
+        violationsInThisFile: list[Violation] = _checkFile(
             filename,
             style=style,
             argTypeHintsInSignature=argTypeHintsInSignature,
@@ -611,7 +612,7 @@ def _checkFile(
         treatPropertyMethodsAsClassAttributes: bool = False,
         requireReturnSectionWhenReturningNothing: bool = False,
         requireYieldSectionWhenYieldingNothing: bool = False,
-) -> List[Violation]:
+) -> list[Violation]:
     if not filename.is_file():  # sometimes folder names can end with `.py`
         return []