+++ pydoclint/parse_config.py
@@ -1,7 +1,9 @@
+from __future__ import annotations
+
 import logging
 import sys
 from pathlib import Path
-from typing import Any, Dict, Optional, Sequence
+from typing import Any, Sequence
 
 import click
 
@@ -14,8 +16,8 @@ else:
 def injectDefaultOptionsFromUserSpecifiedTomlFilePath(
         ctx: click.Context,
         param: click.Parameter,
-        value: Optional[str],
-) -> Optional[str]:
+        value: str | None,
+) -> str | None:
     """
     Inject default objects from user-specified .toml file path.
 
@@ -25,13 +27,13 @@ def injectDefaultOptionsFromUserSpecifiedTomlFilePath(
         The "click" context
     param : click.Parameter
         The "click" parameter; not used in this function; just a placeholder
-    value : Optional[str]
+    value : str | None
         The full path of the .toml file. (It needs to be named ``value``
         so that ``click`` can correctly use it as a callback function.)
 
     Returns
     -------
-    Optional[str]
+    str | None
         The full path of the .toml file
     """
     if not value:
@@ -43,7 +45,7 @@ def injectDefaultOptionsFromUserSpecifiedTomlFilePath(
     return value
 
 
-def parseToml(paths: Optional[Sequence[str]]) -> Dict[str, Any]:
+def parseToml(paths: Sequence[str] | None) -> dict[str, Any]:
     """Parse the pyproject.toml located in the common parent of ``paths``"""
     if paths is None:
         return {}
@@ -56,7 +58,7 @@ def parseToml(paths: Optional[Sequence[str]]) -> Dict[str, Any]:
     return parseOneTomlFile(tomlFilename)
 
 
-def parseOneTomlFile(tomlFilename: Path) -> Dict[str, Any]:
+def parseOneTomlFile(tomlFilename: Path) -> dict[str, Any]:
     """Parse a .toml file"""
     if not tomlFilename.exists():
         logging.info(f'File "{tomlFilename}" does not exist; nothing to load.')
@@ -105,9 +107,9 @@ def findCommonParentFolder(
     return common_parent
 
 
-def updateCtxDefaultMap(ctx: click.Context, config: Dict[str, Any]) -> None:
+def updateCtxDefaultMap(ctx: click.Context, config: dict[str, Any]) -> None:
     """Update the ``click`` context default map with the provided ``config``"""
-    default_map: Dict[str, Any] = {}
+    default_map: dict[str, Any] = {}
     if ctx.default_map:
         default_map.update(ctx.default_map)