+++ pydoclint/utils/astTypes.py
@@ -1,7 +1,11 @@
+from __future__ import annotations
+
 import ast
 import sys
 from typing import Union
 
+# typing.Union is still needed when defining custom types in Python 3.9.
+# It can be changed to `xxx | yyy` after Python 3.9 is dropped.
 FuncOrAsyncFuncDef = Union[ast.AsyncFunctionDef, ast.FunctionDef]
 ClassOrFunctionDef = Union[ast.ClassDef, ast.AsyncFunctionDef, ast.FunctionDef]